Polish `RequestQueue`
By extracting common thread and queueing logic into `EventLoop<T>`.
